+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.7] - 2026-04-22
+
+### Changed
+
+- Default server group to `main-server-group` in topology YAML files when
+  `group` is not specified
+
 ## [0.4.6] - 2026-04-22
 
 ### Fixed
@@ -353,6 +360,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - First release 🎉
+
+[0.4.7]: https://github.com/hpehl/wado/compare/v0.4.6...v0.4.7
 
 [0.4.6]: https://github.com/hpehl/wado/compare/v0.4.5...v0.4.6